+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0__03
 - 修改报错日志问题
+- 添加blog的article和article-category的schemas目录
 
 ## 1.1.0__02
 - 修改登录日志